DistanceService refactor

+++ src/ApiDePapas.Application/Interfaces/IDistanceService.cs
@@ -6,6 +6,6 @@ namespace ApiDePapas.Application.Interfaces
         /// Devuelve la distancia (km) entre dos CPA (códigos postales argentinos).
         /// Debe aceptar CPAs completos (ej: "H3500ABC") o al menos su primera letra.
         /// </summary>
-        public Task<double> GetDistanceKm(string originCpa, string destinationCpa);
+        public Task<float> GetDistanceKm(string originCpa, string destinationCpa);
     }
 }
